rework  interval drawing
